+++ packages/ossa/src/components/modal/index.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-import { View, Image, Text } from "@tarojs/components";
+import { View, Image, Text, Button } from "@tarojs/components";
 import classNames from "classnames";
-//引入组件对应的 类型文件 .d.ts
 import { OsModalProps } from "../../../types/index";
+import getOpenTypeRelatedProps from "../../utils/getOpenTypeRelatedProps";
 
 const closeIconImg =
   "https://yanxuan-static.nosdn.127.net/hxm/yanxuan-wap/p/20161201/style/img/icon-normal/modalClose-9d2d6d39f7.png?imageView";
@@ -20,9 +20,7 @@ function getClassObject(props: OsModalProps) {
 }
 
 function onClose(props: OsModalProps) {
-  if (props.onClose) {
-    props.onClose();
-  }
+  props.onClose?.();
 }
 
 function onClickMask(props: OsModalProps) {
@@ -35,9 +33,7 @@ function onClickCancelBtn(props: OsModalProps) {
   if (props.disableCancelBtn) {
     return;
   }
-  if (props.onCancel) {
-    props.onCancel();
-  }
+  props.onCancel?.();
   onClose(props);
 }
 
@@ -45,10 +41,8 @@ function onClickConfirmBtn(props: OsModalProps) {
   if (props.disableConfirmBtn) {
     return;
   }
-  if (props.onConfirm) {
-    props.onConfirm();
-  }
-  onClose(props);
+  props.onConfirm?.()
+  props.closeOnConfirm && onClose(props);
 }
 
 export default function Index(props: OsModalProps) {
@@ -56,6 +50,7 @@ export default function Index(props: OsModalProps) {
   const classObject = getClassObject(props); //组件修饰
   const styleObject = getStyleObj(props);
   const className = classNames(rootClassName, classObject, props.className);
+  const openTypeRelatedProps = getOpenTypeRelatedProps(props)
 
   const {
     title,
@@ -66,6 +61,7 @@ export default function Index(props: OsModalProps) {
     showCloseIcon = true,
     closable = true,
     closeIconPosition,
+    confirmOpenType
   } = props;
 
   const isRenderAction = confirmText || cancelText;
@@ -102,7 +98,7 @@ export default function Index(props: OsModalProps) {
           {isRenderAction && (
             <View className='ossa-modal__footer'>
               {cancelText && (
-                <View
+                <Button
                   className={classNames({
                     ["ossa-action-btn"]: true,
                     ["ossa-action-btn--cancel"]: true,
@@ -111,19 +107,21 @@ export default function Index(props: OsModalProps) {
                   onClick={(e) => onClickCancelBtn(props)}
                 >
                   {cancelText}
-                </View>
+                </Button>
               )}
               {confirmText && (
-                <View
+                <Button
                   className={classNames({
                     ["ossa-action-btn"]: true,
                     ["ossa-action-btn--confirm"]: true,
                     ["ossa-action-btn--disabled"]: props.disableConfirmBtn,
                   })}
+                  openType={confirmOpenType}
+                  {...openTypeRelatedProps}
                   onClick={(e) => onClickConfirmBtn(props)}
                 >
                   {confirmText}
-                </View>
+                </Button>
               )}
             </View>
           )}
@@ -149,7 +147,7 @@ export default function Index(props: OsModalProps) {
           {isRenderAction && (
             <View className='ossa-modal__footer'>
               {cancelText && (
-                <View
+                <Button
                   className={classNames({
                     ["ossa-action-btn"]: true,
                     ["ossa-action-btn--cancel"]: true,
@@ -158,19 +156,21 @@ export default function Index(props: OsModalProps) {
                   onClick={(e) => onClickCancelBtn(props)}
                 >
                   {cancelText}
-                </View>
+                </Button>
               )}
               {confirmText && (
-                <View
+                <Button
                   className={classNames({
                     ["ossa-action-btn"]: true,
                     ["ossa-action-btn--confirm"]: true,
                     ["ossa-action-btn--disabled"]: props.disableConfirmBtn,
                   })}
+                  openType={confirmOpenType}
+                  {...openTypeRelatedProps}
                   onClick={(e) => onClickConfirmBtn(props)}
                 >
                   {confirmText}
-                </View>
+                </Button>
               )}
             </View>
           )}
@@ -187,6 +187,7 @@ Index.defaultProps = {
   cancelText: null,
   confirmText: null,
   closeOnClickMask: true,
+  closeOnConfirm: false,
   showCloseIcon: true,
   closeIconPosition: "top-right",
   onClose: null,